inokit/requirements.txt
Django>=4.2.0
Pillow>=10.0.0
django-widget-tweaks>=1.4.12
